+++ packages/agtree/CHANGELOG.md
@@ -8,6 +8,13 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## Unreleased
+
+### Fixed
+
+- Fixed missing children data in the deserializer for certain nodes.
+- Add error messages for rules parsing errors.
+
 ## [2.1.3] - 2024-10-21
 
 ### Fixed